Refactor handlers, middleware, models, and json functions to api module
This change was made to organize the project by decluttering the project's
root directory. main.go was also moved to the cmd/gorage-sale folder for
the same reason. The project will now be able to incorporate htmx and
tailwind css on the front end in an organized manner.
