+++ internal/api/handle_oauth_google_login.go
@@ -1,11 +1,11 @@
-package main
+package api
 
 import (
 	"github.com/avearmin/gorage-sale/internal/oauth2"
 	"net/http"
 )
 
-func (cfg apiConfig) handleOAuthGoogleLogin(w http.ResponseWriter, r *http.Request) {
+func (cfg config) HandleOAuthGoogleLogin(w http.ResponseWriter, r *http.Request) {
 	switch r.Method {
 	case http.MethodGet:
 		cfg.getOAuthGoogleLogin(w, r)
@@ -14,7 +14,7 @@ func (cfg apiConfig) handleOAuthGoogleLogin(w http.ResponseWriter, r *http.Reque
 	}
 }
 
-func (cfg apiConfig) getOAuthGoogleLogin(w http.ResponseWriter, r *http.Request) {
+func (cfg config) getOAuthGoogleLogin(w http.ResponseWriter, r *http.Request) {
 	oauthState, err := cfg.StateStore.GenerateState()
 	if err != nil {
 		respondWithError(w, http.StatusInternalServerError, "error generating state")
