+++ internal/api/handle_users_test.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import "testing"
