+++ internal/api/model_item.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"time"
@@ -6,14 +6,6 @@ import (
 	"github.com/avearmin/gorage-sale/internal/database"
 	"github.com/google/uuid"
 )
-
-type User struct {
-	ID        uuid.UUID `json:"id"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-	Name      string    `json:"name"`
-	Email     string    `json:"email"`
-}
 
 type Item struct {
 	ID          uuid.UUID `json:"id"`
@@ -24,16 +16,6 @@ type Item struct {
 	Price       int32     `json:"price"`
 	Sold        bool      `json:"sold"`
 	SellerID    uuid.UUID `json:"seller_id"`
-}
-
-func dbUserToJSONUser(user database.User) User {
-	return User{
-		ID:        user.ID,
-		CreatedAt: user.CreatedAt,
-		UpdatedAt: user.UpdatedAt,
-		Name:      user.Name,
-		Email:     user.Email,
-	}
 }
 
 func dbItemToJSONItem(item database.Item) Item {
