+++ internal/api/model_user.go
@@ -1,0 +1,26 @@
+package api
+
+import (
+	"time"
+
+	"github.com/avearmin/gorage-sale/internal/database"
+	"github.com/google/uuid"
+)
+
+type User struct {
+	ID        uuid.UUID `json:"id"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+	Name      string    `json:"name"`
+	Email     string    `json:"email"`
+}
+
+func dbUserToJSONUser(user database.User) User {
+	return User{
+		ID:        user.ID,
+		CreatedAt: user.CreatedAt,
+		UpdatedAt: user.UpdatedAt,
+		Name:      user.Name,
+		Email:     user.Email,
+	}
+}
