+++ internal/api/respond_with_error.go
@@ -1,0 +1,19 @@
+package api
+
+import (
+	"log"
+	"net/http"
+)
+
+func respondWithError(w http.ResponseWriter, code int, msg string) {
+	if code/100 == 5 {
+		log.Printf("Responding with status code %d: %s", code, msg)
+	}
+
+	var errorResponse struct {
+		Error string `json:"error"`
+	}
+	errorResponse.Error = msg
+
+	respondWithJSON(w, code, errorResponse)
+}
